Move the tint logic to the theme
